Bug fixes (scoring), more uniprot functionality
Bug fixes and all known memory leaks fixed (profiled via Valgrind).
Major bugfix to scoring system - all tests need to be rerun

+++ is.c
@@ -24,6 +24,7 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
+#include <stdint.h>
 #include <stdlib.h>
 
 #ifdef USE_MALLOC_WRAPPERS
@@ -31,18 +32,18 @@
 #endif
 
 typedef unsigned char ubyte_t;
-#define chr(i) (cs == sizeof(int) ? ((const int *)T)[i]:((const unsigned char *)T)[i])
+#define chr(i) (cs == sizeof(int64_t) ? ((const int64_t *)T)[i]:((const unsigned char *)T)[i])
 
 /* find the start or end of each bucket */
-static void getCounts(const unsigned char *T, int *C, int n, int k, int cs)
+static void getCounts(const unsigned char *T, int64_t *C, int64_t n, int64_t k, int64_t cs)
 {
-	int i;
+	int64_t i;
 	for (i = 0; i < k; ++i) C[i] = 0;
 	for (i = 0; i < n; ++i) ++C[chr(i)];
 }
-static void getBuckets(const int *C, int *B, int k, int end)
+static void getBuckets(const int64_t *C, int64_t *B, int64_t k, int64_t end)
 {
-	int i, sum = 0;
+	int64_t i, sum = 0;
 	if (end) {
 		for (i = 0; i < k; ++i) {
 			sum += C[i];
@@ -57,10 +58,10 @@ static void getBuckets(const int *C, int *B, int k, int end)
 }
 
 /* compute SA */
-static void induceSA(const unsigned char *T, int *SA, int *C, int *B, int n, int k, int cs)
+static void induceSA(const unsigned char *T, int64_t *SA, int64_t *C, int64_t *B, int64_t n, int64_t k, int64_t cs)
 {
-	int *b, i, j;
-	int  c0, c1;
+	int64_t *b, i, j;
+	int64_t  c0, c1;
 	/* compute SAl */
 	if (C == B) getCounts(T, C, n, k, cs);
 	getBuckets(C, B, k, 0);	/* find starts of buckets */
@@ -97,19 +98,19 @@ static void induceSA(const unsigned char *T, int *SA, int *C, int *B, int n, int
  * find the suffix array SA of T[0..n-1] in {0..k-1}^n use a working
  * space (excluding T and SA) of at most 2n+O(1) for a constant alphabet
  */
-static int sais_main(const unsigned char *T, int *SA, int fs, int n, int k, int cs)
+static int64_t sais_main(const unsigned char *T, int64_t *SA, int64_t fs, int64_t n, int64_t k, int64_t cs)
 {
-	int *C, *B, *RA;
-	int  i, j, c, m, p, q, plen, qlen, name;
-	int  c0, c1;
-	int  diff;
+	int64_t *C, *B, *RA;
+	int64_t  i, j, c, m, p, q, plen, qlen, name;
+	int64_t  c0, c1;
+	int64_t  diff;
 
 	/* stage 1: reduce the problem by at least 1/2 sort all the
 	 * S-substrings */
 	if (k <= fs) {
 		C = SA + n;
 		B = (k <= (fs - k)) ? C + k : C;
-	} else if ((C = B = (int *) malloc(k * sizeof(int))) == NULL) return -2;
+	} else if ((C = B = (int64_t *) malloc(k * sizeof(int64_t))) == NULL) return -2;
 	getCounts(T, C, n, k, cs);
 	getBuckets(C, B, k, 1);	/* find ends of buckets */
 	for (i = 0; i < n; ++i) SA[i] = 0;
@@ -156,7 +157,7 @@ static int sais_main(const unsigned char *T, int *SA, int fs, int n, int k, int 
 		for (i = n - 1, j = m - 1; m <= i; --i) {
 			if (SA[i] != 0) RA[j--] = SA[i] - 1;
 		}
-		if (sais_main((unsigned char *) RA, SA, fs + n - m * 2, m, name, sizeof(int)) != 0) return -2;
+		if (sais_main((unsigned char *) RA, SA, fs + n - m * 2, m, name, sizeof(int64_t)) != 0) return -2;
 		for (i = n - 2, j = m - 1, c = 0, c1 = chr(n - 1); 0 <= i; --i, c1 = c0) {
 			if ((c0 = chr(i)) < (c1 + c)) c = 1;
 			else if (c != 0) RA[j--] = i + 1, c = 0; /* get p1 */
@@ -167,7 +168,7 @@ static int sais_main(const unsigned char *T, int *SA, int fs, int n, int k, int 
 	if (k <= fs) {
 		C = SA + n;
 		B = (k <= (fs - k)) ? C + k : C;
-	} else if ((C = B = (int *) malloc(k * sizeof(int))) == NULL) return -2;
+	} else if ((C = B = (int64_t *) malloc(k * sizeof(int64_t))) == NULL) return -2;
 	/* put all left-most S characters into their buckets */
 	getCounts(T, C, n, k, cs);
 	getBuckets(C, B, k, 1);	/* find ends of buckets */
@@ -188,7 +189,7 @@ static int sais_main(const unsigned char *T, int *SA, int fs, int n, int k, int 
  * @param n The length of the given string.
  * @return 0 if no error occurred
  */
-int is_sa(const ubyte_t *T, int *SA, int n)
+int64_t is_sa(const ubyte_t *T, int64_t *SA, int64_t n)
 {
 	if ((T == NULL) || (SA == NULL) || (n < 0)) return -1;
 	SA[0] = n;
@@ -205,10 +206,10 @@ int is_sa(const ubyte_t *T, int *SA, int n)
  * @param n The length of the given string.
  * @return The primary index if no error occurred, -1 or -2 otherwise.
  */
-int is_bwt(ubyte_t *T, int n)
+int64_t is_bwt(ubyte_t *T, int64_t n)
 {
-	int *SA, i, primary = 0;
-	SA = (int*)calloc(n+1, sizeof(int));
+	int64_t *SA, i, primary = 0;
+	SA = (int64_t *)calloc(n+1, sizeof(int64_t));
 
 	if (is_sa(T, SA, n)) return -1;
